# shared/openapi.yaml
openapi: 3.1.0
info:
  title: Capture API
  version: 0.1.0
  description: |
    HTTP contract between the Capture iOS app and the Capture server.
    The iOS networking layer mirrors these shapes by hand today; once the
    contract stabilizes we'll codegen `APITypes.swift` from this file.
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://api.capture.app
    description: Production (placeholder)
paths:
...
  /photographers:
    get:
      summary: Search nearby photographers
      parameters:
        - { name: lat, in: query, required: true, schema: { type: number } }
        - { name: lng, in: query, required: true, schema: { type: number } }
        - { name: radiusKm, in: query, required: false, schema: { type: number, default: 25 } }
        - { name: category, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: Photographer cards
          content:
            application/json:
              schema:
                type: object
                properties:
                  photographers:
                    type: array
                    items: { $ref: '#/components/schemas/PhotographerCard' }
  /photographers/{id}:
    get:
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Photographer detail + services
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PhotographerDetail' }
...
  /me/photographer/onboarding/refresh-link:
    post:
      summary: Mint a fresh Stripe Connect account link (used when prior link expired)
      responses:
        '200':
          description: Fresh account link
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountLinkUrl: { type: string }
                  expiresAt: { type: string, format: date-time }
        '409': { description: Onboarding has not been started }
  /me/photographer:
    get:
      summary: The caller's photographer profile + onboarding status
      responses:
        '200':
          description: Photographer self
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PhotographerMe' }
        '404': { description: Caller is not a photographer }
    patch:
      summary: Update profile basics (bio / hourlyRate / homeCity / timezone / etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bio: { type: string, maxLength: 2000 }
                homeCity: { type: string }
                homeAddress: { type: string }
                heroImageUrl: { type: string, format: uri }
                hourlyRateCents: { type: integer, minimum: 0 }
                timezone: { type: string, description: 'IANA timezone' }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PhotographerMe' }
  /me/photographer/services:
    post:
      summary: Create a new service offering
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ServiceUpsert' }
      responses:
        '201': { description: Created }
  /me/photographer/services/{id}:
    patch:
      summary: Update one of the caller's services
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ServiceUpsert' }
      responses:
        '200': { description: Updated }
        '404': { description: Not found or not owned by caller }
    delete:
      summary: Soft-delete a service (sets active=false)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Deleted }
        '409': { description: Has upcoming bookings — cannot delete }
  /me/photographer/availability/slots:
    post:
      summary: Bulk-create availability slots
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slots]
              properties:
                slots:
                  type: array
                  minItems: 1
                  maxItems: 50
                  items:
                    type: object
                    required: [startsAt, endsAt]
                    properties:
                      startsAt: { type: string, format: date-time }
                      endsAt: { type: string, format: date-time }
      responses:
        '201': { description: Slots created }
        '409': { description: Slot overlaps existing availability }
  /me/photographer/availability/slots/{id}:
    delete:
      summary: Delete an open slot
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Deleted }
        '409': { description: Slot is held or booked }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    PhotographerCard:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        avatarUrl: { type: string, nullable: true }
        heroImageUrl: { type: string, nullable: true }
        homeCity: { type: string }
        bio: { type: string, nullable: true }
        hourlyRateCents: { type: integer }
        avgRating: { type: number }
        ratingCount: { type: integer }
        distanceMeters: { type: number }
    PhotographerDetail:
      allOf:
        - $ref: '#/components/schemas/PhotographerCard'
        - type: object
          properties:
            services:
              type: array
              items: { $ref: '#/components/schemas/Service' }
    Service:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        durationMinutes: { type: integer }
        priceCents: { type: integer }
        category: { type: string }
...
    PhotographerMe:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        bio: { type: string, nullable: true }
        homeCity: { type: string }
        homeAddress: { type: string, nullable: true }
        heroImageUrl: { type: string, nullable: true }
        hourlyRateCents: { type: integer }
        timezone: { type: string }
        stripeAccountId: { type: string, nullable: true }
        onboardingStartedAt: { type: string, format: date-time, nullable: true }
        onboardingCompletedAt: { type: string, format: date-time, nullable: true }
        avgRating: { type: number }
        ratingCount: { type: integer }
    ServiceUpsert:
      type: object
      required: [title, durationMinutes, priceCents, category]
      properties:
        title: { type: string }
        description: { type: string }
        durationMinutes: { type: integer, minimum: 5 }
        priceCents: { type: integer, minimum: 0 }
        category:
          type: string
          enum: [portrait, wedding, wedding_video, event, family, headshot, other]
        active: { type: boolean }
...
security:
  - bearerAuth: []